requirement.txt
appdirs==1.4.3
Django==1.11.3
django-froala-editor==2.6.4
django-jet==1.0.6
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
